Css for market page

+++ jardiquest/model/database/entity/commande.py
@@ -5,7 +5,7 @@ class Commande(db.Model):
     __tablename__ = "commande"
 
     idCommande = db.Column(db.String(100), primary_key=True)
-    quantite = db.Column(db.Integer, nullable=False)
+    quantite = db.Column(db.Float, nullable=False)
     dateAchat = db.Column(db.Date, nullable=False)
     traitee = db.Column(db.Boolean, default=False)
